requirements.txt
pip~=25.2
typing_extensions~=4.10.0
attrs~=22.2.0
PySocks~=1.7.1
protobuf~=5.27.2
cryptography~=41.0.7
filelock~=3.13.1
pyee~=9.0.4
trio~=0.22.2
pytest~=7.4.3
allure-python-commons~=2.13.5
MarkupSafe~=2.1.3
Pygments~=2.16.1
wcwidth~=0.2.8
pluggy~=1.0.0
exceptiongroup~=1.1.1
iniconfig~=2.0.0
py~=1.11.0
pexpect~=4.8.0
numpy~=1.26.3
certifi~=2022.12.7
python-slugify~=6.1.2
urllib3~=1.26.15
greenlet~=2.0.1
psutil~=5.9.5
Pillow~=10.0.1
idna~=3.4
charset-normalizer~=3.1.0
packaging~=23.0
playwright~=1.31.1
setuptools~=65.5.0
trove-classifiers~=2023.11.22
pyparsing~=3.2.3
Jinja2~=3.1.2
build~=1.0.3
matplotlib~=3.10.3
allure-pytest~=2.13.5
pytest-playwright~=0.3.2